finish all except bonus

+++ php/deletemeal.php
@@ -11,6 +11,22 @@
         die("No mealname");
     }
     $mealname = $_POST['mealname'];
+    $shopname = $_POST['shopname'];
+
+    // check if there exist unfinished order containing editing meal
+    $status = "Not Finished";
+    $orderquery = "SELECT price, quantity, mealname FROM i_order WHERE shopname = ? and mealname = ? and status = ?";
+    // $o_stmt = $conn->query($orderquery);
+    $o_stmt = $conn->prepare($orderquery);   // avoid sql injection
+    $o_stmt->bind_param("sss", $shopname, $mealname, $status);  // 's' specifies the variable type => 'string'
+    $o_stmt->execute();
+    $o_result = $o_stmt->get_result();
+    if($o_result->num_rows > 0){
+        die("Cannot Update!! Because there exists unfinished order containing this meal!!");
+    }
+
+
+
     /** check if shop name already exist or not */ 
     // delete from `user` where userid='$id'";
     $sql =  "DELETE FROM meal  WHERE mealname = ?";
